+++ src/pages/homePage/homePage.vue
@@ -13,10 +13,6 @@
 		</div>
 		<div class="content">
 			<div class="swiper-wrapper">
-				<!-- <swiper :options="swiperOption" ref="mySwiper">
-					<swiper-slide v-for="(item,index) in bannerList" :key="index"><img :src="item.img" /></swiper-slide>
-					<div class="swiper-pagination" slot="pagination"></div>
-				</swiper> -->
 				<Swiper class="swiper" :autoPlay="true" v-if="bannerList.length > 0">
 					<Slide v-for="(item,index) in bannerList" :key="index">
 						<img :src="item.img" />
@@ -35,7 +31,7 @@
 					<div class="products">
 						<div class="product-list" v-for="(category,index) in goods" :key="index" ref="productList">
 							<p class="category-title">{{category.name}}</p>
-							<div class="product" v-for="(product,index) in category.product" :key="index" @click="showProductModal(product.id)">
+							<div class="product" v-for="(product,index) in category.products" :key="index" @click="showProductModal(product.id)">
 								<div class="product-img" :style="{background:'url('+product.showImg+') center no-repeat'}"></div>
 								<div class="product-info">
 									<span class="product-name">{{product.name}}</span>
@@ -52,13 +48,6 @@
 					</div>
 				</div>
 			</div>
-			<!-- <div class="cart-wrapper">
-				<span class="cart-pic" @click.native="cartModalShow=true">
-					<img src="../../images/cart.png" />
-					<badge text="8"></badge>
-				</span>
-				<div class="pay-btn" @click="goPay">去结算</div>
-			</div> -->
 		</div>
 		<transition name="swipe-up">
 			<div class="product-mask" v-transfer-dom v-show="productModalShow">
@@ -367,7 +356,7 @@ export default {
 		},
 		getGoodsByShop(shopId) {
 			return  getAllGoods(shopId).then(res=>{
-			    let goods=res.data;
+			    let goods=res;
 			    goods.map(i => {
 				i.products.map(e => {
 					e.showImg = IMG_PATH + e.imgs.split(',')[0];
@@ -375,718 +364,6 @@ export default {
 				});
 				this.goods=goods;
 			});
-
-			// const goods = [
-			// 	{
-			// 		id: 4,
-			// 		shopId: 1,
-			// 		parentId: 0,
-			// 		name: '经典热咖啡系列',
-			// 		isShow: true,
-			// 		sort: 1,
-			// 		dataFlag: 1,
-			// 		createDateTime: '2018-07-22T12:39:50',
-			// 		product: [
-			// 			{
-			// 				id: 20,
-			// 				code: '020',
-			// 				name: '卡布基诺咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:20:49',
-			// 				status: 1,
-			// 				description: '卡布基诺咖啡',
-			// 				price: 10,
-			// 				discount: 0,
-			// 				realPrice: 10,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/kabuqiluo_1.png',
-			// 			},
-			// 			{
-			// 				id: 21,
-			// 				code: '021',
-			// 				name: '美式咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:20:52',
-			// 				status: 1,
-			// 				description: '美式咖啡',
-			// 				price: 8,
-			// 				discount: 0,
-			// 				realPrice: 8,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/bingmeishi_1.png',
-			// 			},
-			// 			{
-			// 				id: 22,
-			// 				code: '022',
-			// 				name: '拿铁咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:20:55',
-			// 				status: 1,
-			// 				description: '拿铁咖啡',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/natie_1.png',
-			// 			},
-			// 			{
-			// 				id: 23,
-			// 				code: '023',
-			// 				name: '意式浓缩咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:20:59',
-			// 				status: 1,
-			// 				description: '意式浓缩咖啡',
-			// 				price: 8,
-			// 				discount: 0,
-			// 				realPrice: 8,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/yishinongsuo_1.png',
-			// 			},
-			// 			{
-			// 				id: 24,
-			// 				code: '024',
-			// 				name: '招牌卡布基诺咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:21:07',
-			// 				status: 1,
-			// 				description: '招牌卡布基诺咖啡',
-			// 				price: 11,
-			// 				discount: 0,
-			// 				realPrice: 11,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/zhaobaikabuqiluoda_1.png',
-			// 			},
-			// 			{
-			// 				id: 25,
-			// 				code: '025',
-			// 				name: '招牌美式咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:21:13',
-			// 				status: 1,
-			// 				description: '招牌美式咖啡',
-			// 				price: 9,
-			// 				discount: 0,
-			// 				realPrice: 9,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/zhaopaimeishi_1.png',
-			// 			},
-			// 			{
-			// 				id: 26,
-			// 				code: '026',
-			// 				name: '招牌拿铁咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:21:20',
-			// 				status: 1,
-			// 				description: '招牌拿铁咖啡',
-			// 				price: 13,
-			// 				discount: 0,
-			// 				realPrice: 13,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/zhaopainatie_1.png',
-			// 			},
-			// 			{
-			// 				id: 27,
-			// 				code: '027',
-			// 				name: '招牌意式浓缩咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 4,
-			// 				createDateTime: '2018-07-22T17:21:27',
-			// 				status: 1,
-			// 				description: '招牌意式浓缩咖啡',
-			// 				price: 9,
-			// 				discount: 0,
-			// 				realPrice: 9,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/rekafei/zhaopaiyisinongsuo_1.png',
-			// 			},
-			// 		],
-			// 	},
-			// 	{
-			// 		id: 3,
-			// 		shopId: 1,
-			// 		parentId: 0,
-			// 		name: '经典冰咖啡系列',
-			// 		isShow: true,
-			// 		sort: 2,
-			// 		dataFlag: 1,
-			// 		createDateTime: '2018-07-22T12:39:32',
-			// 		product: [
-			// 			{
-			// 				id: 14,
-			// 				code: '014',
-			// 				name: '冰卡布基诺咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 3,
-			// 				createDateTime: '2018-07-22T16:58:38',
-			// 				status: 1,
-			// 				description: '冰卡布基诺咖啡',
-			// 				price: 11,
-			// 				discount: 0,
-			// 				realPrice: 11,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/bingkafei/bingkabuqiluokafei_1.png',
-			// 			},
-			// 			{
-			// 				id: 15,
-			// 				code: '015',
-			// 				name: '冰美式咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 3,
-			// 				createDateTime: '2018-07-22T16:58:44',
-			// 				status: 1,
-			// 				description: '冰美式咖啡',
-			// 				price: 9,
-			// 				discount: 0,
-			// 				realPrice: 9,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/bingkafei/bingmeishikafei_1.png',
-			// 			},
-			// 			{
-			// 				id: 16,
-			// 				code: '016',
-			// 				name: '冰拿铁咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 3,
-			// 				createDateTime: '2018-07-22T16:58:50',
-			// 				status: 1,
-			// 				description: '冰拿铁咖啡',
-			// 				price: 13,
-			// 				discount: 0,
-			// 				realPrice: 13,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/bingkafei/bingnatiekafei_1.png',
-			// 			},
-			// 			{
-			// 				id: 17,
-			// 				code: '017',
-			// 				name: '招牌冰卡布基诺咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 3,
-			// 				createDateTime: '2018-07-22T16:59:06',
-			// 				status: 1,
-			// 				description: '招牌冰卡布基诺咖啡',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/bingkafei/zhaopaibingkabuqiluo_1.png',
-			// 			},
-			// 			{
-			// 				id: 18,
-			// 				code: '018',
-			// 				name: '招牌冰美式咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 3,
-			// 				createDateTime: '2018-07-22T16:59:11',
-			// 				status: 1,
-			// 				description: '招牌冰美式咖啡',
-			// 				price: 10,
-			// 				discount: 0,
-			// 				realPrice: 10,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/bingkafei/zhaopaibingmeishi_1.png',
-			// 			},
-			// 			{
-			// 				id: 19,
-			// 				code: '019',
-			// 				name: '招牌冰拿铁咖啡',
-			// 				shopId: 1,
-			// 				goodsCatsId: 3,
-			// 				createDateTime: '2018-07-22T16:59:15',
-			// 				status: 1,
-			// 				description: '招牌冰拿铁咖啡',
-			// 				price: 14,
-			// 				discount: 0,
-			// 				realPrice: 14,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/bingkafei/zhaopaibingnatie_1.png',
-			// 			},
-			// 		],
-			// 	},
-			// 	{
-			// 		id: 5,
-			// 		shopId: 1,
-			// 		parentId: 0,
-			// 		name: '咖啡机鲜萃茶系列',
-			// 		isShow: true,
-			// 		sort: 3,
-			// 		dataFlag: 1,
-			// 		createDateTime: '2018-07-22T12:40:09',
-			// 		product: [
-			// 			{
-			// 				id: 28,
-			// 				code: '028',
-			// 				name: '阿里山冰茶',
-			// 				shopId: 1,
-			// 				goodsCatsId: 5,
-			// 				createDateTime: '2018-07-22T17:28:28',
-			// 				status: 1,
-			// 				description: '阿里山冰茶',
-			// 				price: 10,
-			// 				discount: 0,
-			// 				realPrice: 10,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/cha/alishanbingcha_1.png',
-			// 			},
-			// 			{
-			// 				id: 29,
-			// 				code: '029',
-			// 				name: '茉莉绿茶',
-			// 				shopId: 1,
-			// 				goodsCatsId: 5,
-			// 				createDateTime: '2018-07-22T17:28:54',
-			// 				status: 1,
-			// 				description: '茉莉绿茶',
-			// 				price: 10,
-			// 				discount: 0,
-			// 				realPrice: 10,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/cha/molilvcha_1.png',
-			// 			},
-			// 			{
-			// 				id: 30,
-			// 				code: '030',
-			// 				name: '炭焙乌龙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 5,
-			// 				createDateTime: '2018-07-22T17:28:58',
-			// 				status: 1,
-			// 				description: '炭焙乌龙',
-			// 				price: 10,
-			// 				discount: 0,
-			// 				realPrice: 10,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/cha/tanbeiwulong_1.png',
-			// 			},
-			// 			{
-			// 				id: 31,
-			// 				code: '031',
-			// 				name: '乌龙青茶',
-			// 				shopId: 1,
-			// 				goodsCatsId: 5,
-			// 				createDateTime: '2018-07-22T17:29:01',
-			// 				status: 1,
-			// 				description: '乌龙青茶',
-			// 				price: 10,
-			// 				discount: 0,
-			// 				realPrice: 10,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/cha/wulongqingcha_1.png',
-			// 			},
-			// 		],
-			// 	},
-			// 	{
-			// 		id: 1,
-			// 		shopId: 1,
-			// 		parentId: 0,
-			// 		name: '冰沙系列',
-			// 		isShow: true,
-			// 		sort: 4,
-			// 		dataFlag: 1,
-			// 		createDateTime: '2018-07-17T22:59:07',
-			// 		product: [
-			// 			{
-			// 				id: 1,
-			// 				code: '001',
-			// 				name: '百香果冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-17T22:55:57',
-			// 				status: 1,
-			// 				description: '百香果沙冰',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/baixiangbingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 2,
-			// 				code: '002',
-			// 				name: '草莓冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-18T23:03:02',
-			// 				status: 1,
-			// 				description: '草莓冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/caomeibingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 3,
-			// 				code: '003',
-			// 				name: '凤梨冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-22T16:40:16',
-			// 				status: 1,
-			// 				description: '凤梨冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/fenglibingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 4,
-			// 				code: '004',
-			// 				name: '红豆牛奶冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-22T16:43:03',
-			// 				status: 1,
-			// 				description: '红豆牛奶冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/hongdouniunaibingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 5,
-			// 				code: '005',
-			// 				name: '柳橙冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-22T16:43:07',
-			// 				status: 1,
-			// 				description: '柳橙冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/liuchengbingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 6,
-			// 				code: '006',
-			// 				name: '绿豆牛奶冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-22T16:43:10',
-			// 				status: 1,
-			// 				description: '绿豆牛奶冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/lvdouniunaibingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 7,
-			// 				code: '007',
-			// 				name: '芒果冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-22T16:43:14',
-			// 				status: 1,
-			// 				description: '芒果冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/mangguobingsha_1.png',
-			// 			},
-			// 			{
-			// 				id: 8,
-			// 				code: '008',
-			// 				name: '奇异果冰沙',
-			// 				shopId: 1,
-			// 				goodsCatsId: 1,
-			// 				createDateTime: '2018-07-22T16:43:17',
-			// 				status: 1,
-			// 				description: '奇异果冰沙',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/qiyiguobingsha_1.png',
-			// 			},
-			// 		],
-			// 	},
-			// 	{
-			// 		id: 6,
-			// 		shopId: 1,
-			// 		parentId: 0,
-			// 		name: '苏打水系列',
-			// 		isShow: true,
-			// 		sort: 5,
-			// 		dataFlag: 1,
-			// 		createDateTime: '2018-07-22T12:40:27',
-			// 		product: [
-			// 			{
-			// 				id: 32,
-			// 				code: '032',
-			// 				name: '百香果苏打水',
-			// 				shopId: 1,
-			// 				goodsCatsId: 6,
-			// 				createDateTime: '2018-07-22T17:35:24',
-			// 				status: 1,
-			// 				description: '百香果苏打水',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/sudashui/baixiangguo_1.png',
-			// 			},
-			// 			{
-			// 				id: 33,
-			// 				code: '033',
-			// 				name: '草莓苏打水',
-			// 				shopId: 1,
-			// 				goodsCatsId: 6,
-			// 				createDateTime: '2018-07-22T17:35:27',
-			// 				status: 1,
-			// 				description: '草莓苏打水',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/sudashui/caomei_1.png',
-			// 			},
-			// 			{
-			// 				id: 34,
-			// 				code: '034',
-			// 				name: '蓝橘苏打水',
-			// 				shopId: 1,
-			// 				goodsCatsId: 6,
-			// 				createDateTime: '2018-07-22T17:35:31',
-			// 				status: 1,
-			// 				description: '蓝橘苏打水',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/sudashui/lanju_1.png',
-			// 			},
-			// 			{
-			// 				id: 35,
-			// 				code: '035',
-			// 				name: '芒果苏打水',
-			// 				shopId: 1,
-			// 				goodsCatsId: 6,
-			// 				createDateTime: '2018-07-22T17:35:35',
-			// 				status: 1,
-			// 				description: '芒果苏打水',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/sudashui/mangguo_1.png',
-			// 			},
-			// 			{
-			// 				id: 36,
-			// 				code: '036',
-			// 				name: '青柠苏打水',
-			// 				shopId: 1,
-			// 				goodsCatsId: 6,
-			// 				createDateTime: '2018-07-22T17:35:39',
-			// 				status: 1,
-			// 				description: '青柠苏打水',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/sudashui/qingning_1.png',
-			// 			},
-			// 			{
-			// 				id: 37,
-			// 				code: '037',
-			// 				name: '水蜜桃苏打',
-			// 				shopId: 1,
-			// 				goodsCatsId: 6,
-			// 				createDateTime: '2018-07-22T17:35:42',
-			// 				status: 1,
-			// 				description: '水蜜桃苏打',
-			// 				price: 12,
-			// 				discount: 0,
-			// 				realPrice: 12,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/sudashui/shuimitao_1.png',
-			// 			},
-			// 		],
-			// 	},
-			// 	{
-			// 		id: 2,
-			// 		shopId: 1,
-			// 		parentId: 0,
-			// 		name: '糕点系列',
-			// 		isShow: true,
-			// 		sort: 6,
-			// 		dataFlag: 1,
-			// 		createDateTime: '2018-07-19T22:56:05',
-			// 		product: [
-			// 			{
-			// 				id: 9,
-			// 				code: '009',
-			// 				name: '草莓QQ松饼',
-			// 				shopId: 1,
-			// 				goodsCatsId: 2,
-			// 				createDateTime: '2018-07-22T16:46:32',
-			// 				status: 1,
-			// 				description: '草莓QQ松饼',
-			// 				price: 7,
-			// 				discount: 0,
-			// 				realPrice: 7,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/gaodian/caomeiqqsongbing_1.png',
-			// 			},
-			// 			{
-			// 				id: 10,
-			// 				code: '010',
-			// 				name: '芒果QQ松饼',
-			// 				shopId: 1,
-			// 				goodsCatsId: 2,
-			// 				createDateTime: '2018-07-22T16:48:18',
-			// 				status: 1,
-			// 				description: '芒果QQ松饼',
-			// 				price: 7,
-			// 				discount: 0,
-			// 				realPrice: 7,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/gaodian/mangguoqqsongbing_1.png',
-			// 			},
-			// 			{
-			// 				id: 11,
-			// 				code: '011',
-			// 				name: '巧克力玛芬蛋糕',
-			// 				shopId: 1,
-			// 				goodsCatsId: 2,
-			// 				createDateTime: '2018-07-22T16:48:23',
-			// 				status: 1,
-			// 				description: '巧克力玛芬蛋糕',
-			// 				price: 6,
-			// 				discount: 0,
-			// 				realPrice: 6,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/gaodian/qiaokelimafengdangao_1.png',
-			// 			},
-			// 			{
-			// 				id: 12,
-			// 				code: '012',
-			// 				name: '原味QQ松饼',
-			// 				shopId: 1,
-			// 				goodsCatsId: 2,
-			// 				createDateTime: '2018-07-22T16:49:16',
-			// 				status: 1,
-			// 				description: '原味QQ松饼',
-			// 				price: 6,
-			// 				discount: 0,
-			// 				realPrice: 6,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/gaodian/yuanweimafengdagao_1.png',
-			// 			},
-			// 			{
-			// 				id: 13,
-			// 				code: '013',
-			// 				name: '原味玛芬蛋糕',
-			// 				shopId: 1,
-			// 				goodsCatsId: 2,
-			// 				createDateTime: '2018-07-22T16:49:19',
-			// 				status: 1,
-			// 				description: '原味玛芬蛋糕',
-			// 				price: 5,
-			// 				discount: 0,
-			// 				realPrice: 5,
-			// 				saleNum: 0,
-			// 				visitNum: 0,
-			// 				dataFlag: 1,
-			// 				imgs: '/oleccoffee-test/gaodian/yuanweiqqsongbing.png',
-			// 			},
-			// 		],
-			// 	},
-			// ];
-			// goods.map(i => {
-			// 	i.product.map(e => {
-			// 		e.showImg = IMG_PATH + e.imgs.split(',')[0];
-			// 	});
-			// });
-			// this.goods = goods;
-
-			// return Promise.resolve(goods);
 		},
 		_initSroll() {
 			this.categoryScroll = new BScroll('.category-wrapper', {
