+++ src/pages/orderConfirm/OrderConfirm.vue
@@ -308,7 +308,7 @@ export default {
 				{
 					appId: data.appId, //公众号名称，由商户传入
 					timeStamp: data.timeStamp, //时间戳，自1970年以来的秒数
-					nonceStr: data.nonceStr, //随机串
+					nonceStr: data.nonceStr, //随机串fdelivery
 					package: data.packageValue, // 统一支付接口返回的prepay_id参数值，提交格式如：prepay_id=\*\*\*）
 					signType: data.signType, //微信签名方式：
 					paySign: data.paySign, //微信签名
